+++ crates/c-api/tests/CMakeLists.txt
@@ -9,20 +9,20 @@ FetchContent_MakeAvailable(googletest)
 
 include(GoogleTest)
 
-function(add_capi_test name path)
-  add_executable(test-${name} ${path})
+function(add_capi_test name)
+  cmake_parse_arguments(PARSE_ARGV 1 arg "" "" "FILES")
+  add_executable(test-${name} ${arg_FILES})
   target_link_libraries(test-${name} PRIVATE wasmtime-cpp gtest_main)
   gtest_discover_tests(test-${name})
 endfunction()
 
-add_capi_test(simple simple.cc)
-add_capi_test(types types.cc)
-add_capi_test(func func.cc)
-add_capi_test(component-instantiate component/instantiate.cc)
+add_capi_test(simple FILES simple.cc)
+add_capi_test(types FILES types.cc)
+add_capi_test(func FILES func.cc)
+add_capi_test(component-instantiate FILES component/instantiate.cc)
+add_capi_test(error FILES error.cc)
 
 # Add a custom test where two files include `wasmtime.hh` and are compiled into
 # the same executable (basically makes sure any defined functions in the header
 # are tagged with `inline`).
-add_executable(test-double-include double-include-a.cc double-include-b.cc)
-target_link_libraries(test-double-include PRIVATE wasmtime-cpp gtest_main)
-gtest_discover_tests(test-double-include)
+add_capi_test(test-double-include FILES double-include-a.cc double-include-b.cc)
